+++ src/lib.rs
@@ -6,3 +6,4 @@
 pub mod converter;
 pub mod event;
 pub mod format;
+pub mod io;